requirements.txt
google-generativeai
python-dotenv
Pillow
openpyxl
streamlit
pandas
